fix #68

+++ backend/routes/api/notesFolder.js
@@ -26,6 +26,7 @@ export async function notesFolder(req) {
 function createNotesFolder(body, userId) {
     // body : {location, name (without directory)}
     if(!body.hasOwnProperty("location") || !body.hasOwnProperty("name")) return new Response("400 Bad Request", {status: 400})
+    if(body.location.includes("..")) return new Response("403 Forbidden", {status: 403})
 
     fs.mkdirSync(`${__dirname}/../../userFiles/${userId}/notes/${body.location}/${body.name}`)
     return new Response("200 Success")
@@ -34,6 +35,7 @@ function createNotesFolder(body, userId) {
 function modifyNotesFolder(body, userId) {
     // body : {location (folder to modify), name (new name )}
     if(!body.hasOwnProperty("location") || !body.hasOwnProperty("name")) return new Response("400 Bad Request", {status: 400})
+    if(body.location.includes("..") || body.name.includes("..")) return new Response("403 Forbidden", {status: 403})
 
     let newLocation = body.location.split("/").slice(0,-1).join("/") + "/" + body.name;
     fs.renameSync(`${__dirname}/../../userFiles/${userId}/notes/${body.location}`, `${__dirname}/../../userFiles/${userId}/notes/${newLocation}`)
@@ -43,6 +45,7 @@ function modifyNotesFolder(body, userId) {
 function deleteNotesFolder(body, userId) {
     // body : {location (folder to delete)}
     if(!body.hasOwnProperty("location")) return new Response("400 Bad Request", {status: 400})
+    if(body.location.includes("..")) return new Response("403 Forbidden", {status: 403})
 
     fs.rmSync(`${__dirname}/../../userFiles/${userId}/notes/${body.location}`,  { recursive: true, force: true })
     return new Response("200 Success")
